vista de boletos comprados

+++ gestor-cines-api-rest/src/main/java/com/ronaldo/gestor/cines/api/rest/models/compraBoletos/CompraBoletos.java
@@ -56,7 +56,7 @@ public class CompraBoletos implements Editable {
               return costoTotal;
        }
 
-       public void setCostoTotal(int costoTotal) {
+       public void setCostoTotal(double costoTotal) {
               this.costoTotal = costoTotal;
        }
